chore: centralize `HTTPClient` creation
This patch aims to avoid code duplication for the HTTPClient initialization and to ensure that we always initialize a client that is capable of both HTTP and HTTPS requests
